+++ weave/panels_py/__init__.py
@@ -1,2 +1,3 @@
 from . import panel_autoboard
 from . import panel_seedboard
+from . import generator_templates
